add server error handling

+++ Application/Activities/DeleteActivity.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -8,12 +9,12 @@ namespace Application.Activities
 {
     public class DeleteActivity
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -21,14 +22,14 @@ namespace Application.Activities
                 _context = context;
 
             }
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
+                if(activity==null) return Result<Unit>.Failure("Not Found");
                 _context.Activities.Remove(activity);
-
-                await _context.SaveChangesAsync();
-
-                return Unit.Value;
+                var result = await _context.SaveChangesAsync()>0;
+                if(!result) return Result<Unit>.Failure("Failed to delete actvity");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
